add license refactor definitons
